# Write out 4 structural indexes at a time instead of 8