I genericized the code that unrolls the index writing loop. When I did this, and told it to write 4 indexes at a time, unrolling the first 8 batches (up to 16 indexes) before doing a loop, I got a roughly 15% speed improvement for stage 1 on my Haswell box!

(This was on clang 16, BTW, and reading twitter.json.)