@lemire 2023-08-24
@jkeiser 

On my todo to review. It looks promising.

There is a standard compliance issue in this PR.

We might want something 'like' this:

```
#if __cpp_constexpr >= 201304
#define SIMDJSON_CONSTEXPR14 constexpr
#else
#define SIMDJSON_CONSTEXPR14
#endif
```