@lemire 2023-08-24
PR https://github.com/simdjson/simdjson/pull/2053 should fix compatibility with pre-C++17 compilers.