@lemire 2023-08-24
GCC 12 on an Ice Lake processor. The number of instructions and the CPU frequency appears constant.

Cycles per byte (stage 1):

| kernel         | main branch | this PR | gain 
|--------------|:-----:|-----------:|--------:|
| Westmere |  0.8752 |       0.8417  |  4% |
| Haswell     |  0.5897 |          0.5333 | 9.5% |
| Ice Lake     | 0.3188  |         0.3206 |  -0.5% |


Methodology: I am using commands like `./build/benchmark/dom/parse jsonexamples/twitter.json -n 10000 -a westmere`.

It seems highly beneficial to haswell. I did not think that a nearly 10% speed boost was possible.