@lemire 2023-08-24
On my macbook air (M2, LLVM 14), I get...

main branch...

```
|- Stage 1
|    Speed        :   9.7495 ns per block ( 52.48%) -   0.1523 ns per byte -   1.7409 ns per structural -   6.5641 GB/s
```

PR:

```
|- Stage 1
|    Speed        :   9.4919 ns per block ( 40.57%) -   0.1483 ns per byte -   1.6949 ns per structural -   6.7422 GB/s
```

So a 2-3% gain which is nice.