@jkeiser 2023-08-24
I just ran with different step sizes from 1-8 on icelake. The instruction count was identical for all. Cycles and branch misses were essentially unchanged.

I am mystified as to where the improvement is coming from, even if it's undeniable.