@lemire 2023-08-25
@jkeiser Keep in mind that you can test different kernels:

```
./build/benchmark/dom/parse jsonexamples/twitter.json -n 10000 -a westmere.
./build/benchmark/dom/parse jsonexamples/twitter.json -n 10000 -a haswell.
./build/benchmark/dom/parse jsonexamples/twitter.json -n 10000 -a icelake.
```

For icelake, we should not see any difference from this PR because it has its own bit extraction routine. And I only saw some tiny blip which was probably spurious.