@jkeiser 2023-08-25
@lemire I didn't realize we'd actually swapped the index writer out, good deal :) Here's the haswell data:

```
Step = 1:

|- Stage 1
|    Speed        :  11.1014 ns per block ( 77.83%) -   0.1735 ns per byte -   1.9824 ns per structural -   5.7647 GB/s
|    Cycles       :  35.4814 per block    ( 91.96%) -   0.5544 per byte    -   6.3358 per structural    -    3.196 GHz est. frequency
|    Instructions : 135.6106 per block    (100.00%) -   2.1190 per byte    -  24.2156 per structural    -    3.822 per cycle
|    Misses       :    1257 branch misses ( 52.64%) - 0 cache misses (  0.00%) - 11315.00 cache references

Step = 2:

|- Stage 1
|    Speed        :  10.1917 ns per block ( 80.23%) -   0.1593 ns per byte -   1.8199 ns per structural -   6.2792 GB/s
|    Cycles       :  32.5754 per block    ( 96.72%) -   0.5090 per byte    -   5.8169 per structural    -    3.196 GHz est. frequency
|    Instructions : 136.5278 per block    (100.00%) -   2.1334 per byte    -  24.3794 per structural    -    4.191 per cycle
|    Misses       :      84 branch misses ( 20.80%) - 0 cache misses (  0.00%) - 11045.00 cache references

Step = 3:

|- Stage 1
|    Speed        :  10.3195 ns per block ( 78.66%) -   0.1613 ns per byte -   1.8427 ns per structural -   6.2015 GB/s
|    Cycles       :  32.9900 per block    ( 98.38%) -   0.5155 per byte    -   5.8909 per structural    -    3.197 GHz est. frequency
|    Instructions : 137.8654 per block    (100.00%) -   2.1543 per byte    -  24.6183 per structural    -    4.179 per cycle
|    Misses       :      51 branch misses ( 75.16%) - 0 cache misses (  0.00%) - 11036.00 cache references

Step = 4:

|- Stage 1
|    Speed        :  10.5268 ns per block ( 85.00%) -   0.1645 ns per byte -   1.8797 ns per structural -   6.0794 GB/s
|    Cycles       :  33.6509 per block    ( 99.07%) -   0.5258 per byte    -   6.0090 per structural    -    3.197 GHz est. frequency
|    Instructions : 136.3611 per block    (100.00%) -   2.1308 per byte    -  24.3497 per structural    -    4.052 per cycle
|    Misses       :      35 branch misses ( 82.23%) - 0 cache misses (  0.00%) - 10696.00 cache references

Step = 5:

|- Stage 1
|    Speed        :  10.5624 ns per block ( 98.65%) -   0.1650 ns per byte -   1.8861 ns per structural -   6.0589 GB/s
|    Cycles       :  33.7647 per block    ( 99.06%) -   0.5276 per byte    -   6.0293 per structural    -    3.197 GHz est. frequency
|    Instructions : 139.0398 per block    (100.00%) -   2.1726 per byte    -  24.8280 per structural    -    4.118 per cycle
|    Misses       :      27 branch misses ( 69.40%) - 0 cache misses (  0.00%) - 11353.00 cache references

Step = 6:

|- Stage 1
|    Speed        :  11.0680 ns per block ( 83.86%) -   0.1729 ns per byte -   1.9764 ns per structural -   5.7821 GB/s
|    Cycles       :  35.3788 per block    ( 98.78%) -   0.5528 per byte    -   6.3175 per structural    -    3.196 GHz est. frequency
|    Instructions : 141.2939 per block    (100.00%) -   2.2078 per byte    -  25.2305 per structural    -    3.994 per cycle
|    Misses       :      29 branch misses ( 67.02%) - 0 cache misses (  0.00%) - 11121.00 cache references

Step = 7:

|- Stage 1
|    Speed        :  11.0336 ns per block ( 84.71%) -   0.1724 ns per byte -   1.9703 ns per structural -   5.8001 GB/s
|    Cycles       :  35.2663 per block    ( 98.99%) -   0.5511 per byte    -   6.2974 per structural    -    3.196 GHz est. frequency
|    Instructions : 143.6110 per block    (100.00%) -   2.2441 per byte    -  25.6443 per structural    -    4.072 per cycle
|    Misses       :      27 branch misses ( 82.32%) - 0 cache misses (  0.00%) - 11387.00 cache references

Step = 8:

|- Stage 1
|    Speed        :  11.3658 ns per block ( 98.90%) -   0.1776 ns per byte -   2.0296 ns per structural -   5.6306 GB/s
|    Cycles       :  36.3307 per block    ( 99.48%) -   0.5677 per byte    -   6.4875 per structural    -    3.196 GHz est. frequency
|    Instructions : 142.2188 per block    (100.00%) -   2.2223 per byte    -  25.3957 per structural    -    3.915 per cycle
|    Misses       :      26 branch misses ( 85.48%) - 0 cache misses (  0.00%) - 10703.00 cache references
```

One thing this tells us is, a difference of 60-70 branch misses is not enough to overcome the difference in the number of instructions per block. A difference of 1200 branch misses, however, is enough to completely overwhelm it. I don't entirely know how to read the percentages involved with branch misses, but it appears that extra if statements don't really affect things that much even when you do 2 writes per if statement.

I'm not 100% sure why step = 2 and step = 1 are so different, though!