@jkeiser 2023-08-25
Interestingly, at least on this haswell test, step = 8 is worse than step = 1.