@lemire 2023-08-25
I need to check whether AWS has Zen processors.