@lemire 2023-08-25
I have a Zen 2 server. I am going to test it out. Results in a few minutes.