@lemire 2023-08-25
zen 4 really likes simdjson...

Stage 1 is roughly twice as fast compared to the results on zen 3...


Main:
```
jsonexamples/twitter.json
=========================
     9867 blocks -     631515 bytes - 55262 structurals (  8.8 %)
special blocks with: utf8      2284 ( 23.1 %) - escape       598 (  6.1 %) - 0 structurals      1287 ( 13.0 %) - 1+ structurals      8581 ( 87.0 %) - 8+ structurals      3272 ( 33.2 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8      1104 ( 11.2 %) - escape       642 (  6.5 %) - 0 structurals       940 (  9.5 %) - 1+ structurals       940 (  9.5 %) - 8+ structurals      2593 ( 26.3 %) - 16+ structurals         0 (  0.0 %)

All Stages (excluding allocation)
|    Speed        :  14.9606 ns per block ( 98.72%) -   0.2338 ns per byte -   2.6715 ns per structural -   4.2777 GB/s
|- Stage 1
|    Speed        :   4.8966 ns per block ( 32.31%) -   0.0765 ns per byte -   0.8744 ns per structural -  13.0694 GB/s
|- Stage 2
|    Speed        :  10.0335 ns per block ( 66.21%) -   0.1568 ns per byte -   1.7917 ns per structural -   6.3782 GB/s

6773.6 documents parsed per second (best)
```


This PR:
```
jsonexamples/twitter.json
=========================
     9867 blocks -     631515 bytes - 55262 structurals (  8.8 %)
special blocks with: utf8      2284 ( 23.1 %) - escape       598 (  6.1 %) - 0 structurals      1287 ( 13.0 %) - 1+ structurals      8581 ( 87.0 %) - 8+ structurals      3272 ( 33.2 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8      1104 ( 11.2 %) - escape       642 (  6.5 %) - 0 structurals       940 (  9.5 %) - 1+ structurals       940 (  9.5 %) - 8+ structurals      2593 ( 26.3 %) - 16+ structurals         0 (  0.0 %)

All Stages (excluding allocation)
|    Speed        :  14.8704 ns per block ( 98.81%) -   0.2324 ns per byte -   2.6554 ns per structural -   4.3036 GB/s
|- Stage 1
|    Speed        :   4.8724 ns per block ( 32.37%) -   0.0761 ns per byte -   0.8701 ns per structural -  13.1344 GB/s
|- Stage 2
|    Speed        :   9.9839 ns per block ( 66.34%) -   0.1560 ns per byte -   1.7828 ns per structural -   6.4100 GB/s

6814.7 documents parsed per second (best)
```


If I ask the zen 4 server to run as *haswell*...


Main branch...

```
jsonexamples/twitter.json
=========================
     9867 blocks -     631515 bytes - 55262 structurals (  8.8 %)
special blocks with: utf8      2284 ( 23.1 %) - escape       598 (  6.1 %) - 0 structurals      1287 ( 13.0 %) - 1+ structurals      8581 ( 87.0 %) - 8+ structurals      3272 ( 33.2 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8      1104 ( 11.2 %) - escape       642 (  6.5 %) - 0 structurals       940 (  9.5 %) - 1+ structurals       940 (  9.5 %) - 8+ structurals      2593 ( 26.3 %) - 16+ structurals         0 (  0.0 %)

All Stages (excluding allocation)
|    Speed        :  16.4817 ns per block ( 98.93%) -   0.2575 ns per byte -   2.9431 ns per structural -   3.8829 GB/s
|- Stage 1
|    Speed        :   8.1141 ns per block ( 48.71%) -   0.1268 ns per byte -   1.4489 ns per structural -   7.8870 GB/s
|- Stage 2
|    Speed        :   8.3453 ns per block ( 50.09%) -   0.1304 ns per byte -   1.4902 ns per structural -   7.6686 GB/s

6148.5 documents parsed per second (best)
```


This PR 
```
jsonexamples/twitter.json
=========================
     9867 blocks -     631515 bytes - 55262 structurals (  8.8 %)
special blocks with: utf8      2284 ( 23.1 %) - escape       598 (  6.1 %) - 0 structurals      1287 ( 13.0 %) - 1+ structurals      8581 ( 87.0 %) - 8+ structurals      3272 ( 33.2 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8      1104 ( 11.2 %) - escape       642 (  6.5 %) - 0 structurals       940 (  9.5 %) - 1+ structurals       940 (  9.5 %) - 8+ structurals      2593 ( 26.3 %) - 16+ structurals         0 (  0.0 %)

All Stages (excluding allocation)
|    Speed        :  16.2923 ns per block ( 98.86%) -   0.2546 ns per byte -   2.9093 ns per structural -   3.9280 GB/s
|- Stage 1
|    Speed        :   7.9896 ns per block ( 48.48%) -   0.1248 ns per byte -   1.4267 ns per structural -   8.0100 GB/s
|- Stage 2
|    Speed        :   8.2845 ns per block ( 50.27%) -   0.1295 ns per byte -   1.4793 ns per structural -   7.7249 GB/s

6220.0 documents parsed per second (best)
```


So again, we get better performance with this PR on recent x64 processors.